+++ mmaction/datasets/video_dataset.py
@@ -1,7 +1,6 @@
 import os.path as osp
 
 import torch
-from mmcv import load
 from mmcv.utils import print_log
 
 from ..core import mean_class_accuracy, top_k_accuracy
@@ -67,23 +66,6 @@ class VideoDataset(BaseDataset):
                     dict(
                         filename=filename,
                         label=onehot if self.multi_class else label))
-        return video_infos
-
-    def load_json_annotations(self):
-        """Load json annotation file to get video information."""
-        video_infos = load(self.ann_file)
-        num_videos = len(video_infos)
-        # only need to process label here
-        if self.multi_class:
-            assert self.num_classes is not None
-            for i in range(num_videos):
-                onehot = torch.zeros(self.num_classes)
-                onehot[video_infos[i]['label']] = 1.0
-                video_infos[i]['label'] = onehot
-        else:
-            for i in range(num_videos):
-                assert len(video_infos[i]['label']) == 1
-                video_infos[i]['label'] = video_infos[i]['label'][0]
         return video_infos
 
     def evaluate(self,
